+++ lib/src/imgui/im_texture.cpp
@@ -15,8 +15,8 @@ void ImTexture::Deleter::operator()(vk::DescriptorSet descriptor_set) const noex
 ImTexture::ImTexture(NotNull<std::shared_ptr<Texture>> const& texture)
 	: m_texture(texture), m_descriptor_set(vk::DescriptorSet{}, Deleter{m_texture->get_image()->get_render_device().get_device()}) {
 	if constexpr (imgui_v) {
-		auto const cis = m_texture->combined_image_sampler();
-		m_descriptor_set.get() = ImGui_ImplVulkan_AddTexture(cis.sampler, cis.image_view, static_cast<VkImageLayout>(vk::ImageLayout::eShaderReadOnlyOptimal));
+		auto const img = m_texture->get_sampler_image();
+		m_descriptor_set.get() = ImGui_ImplVulkan_AddTexture(img.sampler, img.image_view, static_cast<VkImageLayout>(vk::ImageLayout::eShaderReadOnlyOptimal));
 	}
 }
 } // namespace bave